==> 10_symbionts_GH/01_combine_fungal_guilds.ipynb <==
import pandas as pd
import numpy as np
from sklearn import preprocessing
import matplotlib.pyplot as plt
import seaborn as sns

# DATA

#df = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeQuebecMexico/DATA/ITS/Full/Peuplier_REPRISE_ITS_oct2022.ASV_table_rarefied_11982_dnNA.guilds.txt", sep='\t', header=0)
df = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeAspenGreenhouse/DATA/PeuplierGH_jan2024/ITS/all/PeuplierGH_ITS_jan2024.ASV_table_norarefaction_dnNA.guilds.txt", sep='\t', header=0)

#df = df.set_index('#OTU ID')
df.head()

df.columns.values

## Abundance and taxonomy tables 

columns_its = [ele for ele in df.columns if 'ITS' in ele]
columns_tax = [ele for ele in df.columns if 'ITS' not in ele]
df_abu = df[columns_its]
df_abu = df_abu.set_index(df['#OTU ID'])
df_tax = df[columns_tax]
df_tax = df_tax.set_index('#OTU ID')
print(df_tax.shape)
df_tax.head()

df_tax['group'] = df_tax['Guild'].fillna('Other')
df_tax['group'] = df_tax['group'].apply(lambda x: x if x != "-" else "Other")
df_tax['group'] = df_tax.apply(lambda x: x['group'] if x['Confidence Ranking'] in ["Probable","Highly Probable"] else "Other", axis=1)
df_tax['group'] = df_tax['group'].apply(lambda x: 'Mixed' if '-' in str(x) else str(x))
df_tax.head()

df_guilds = df_tax[['group']]
print(df_guilds.shape)
df_guilds

## Subset soil samples

columns_soil = [ele for ele in df if 'sol' in ele and 'Ctrl' not in ele]
df_abu_soil = df_abu[columns_soil]
print(df_abu_soil.shape)
df_abu_soil.head()

## Metadata

df_meta = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeAspenGreenhouse/DATA/PeuplierGH_jan2024/metadata_ITS.txt", sep="\t", header=0)
df_meta.head(3)

# Sum abundances across guilds

print(df_guilds.shape)
print(df_abu_soil.shape)
df_abu_soil.head()

df_sum = df_abu_soil.groupby(df_guilds['group']).sum()
df_sum.to_csv("guilds_02_read_greenhouse.csv", sep=",", header=True, index=True)

df_sum.head()

dt_sum = df_sum.T
stack_guilds = dt_sum.stack().reset_index().rename(columns={'level_0' : 'sample', 0 : 'count'})
df_meta_summed = pd.merge(stack_guilds, df_meta, left_on = 'sample', right_on = 'SampleID', how='left')
df_meta_summed['long.name'] = df_meta_summed['sample'].apply(lambda x: x.replace('-','.'))
df_meta_summed

df_meta_summed.to_csv("guilds_02_read_greenhouse_meta.csv", sep=",", header=True, index=False)

# Rhizosphere

# Subsetting samples
df.columns.values
columns_sub = [ele for ele in df if 'Rhizo' in ele and 'Ctrl' not in ele]
df_abu_sub = df_abu[columns_sub]
print(df_abu_sub.shape)

# Summing guilds
df_sum = df_abu_sub.groupby(df_guilds['group']).sum()
df_sum
df_sum.to_csv("guilds_02_read_greenhouse_rhizo.csv", sep=",", header=True, index=True)

# Merging with metadata
df_meta = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeAspenGreenhouse/DATA/PeuplierGH_jan2024/metadata_ITS.txt", sep="\t", header=0)
dt_sum = df_sum.T
stack_guilds = dt_sum.stack().reset_index().rename(columns={'level_0' : 'sample', 0 : 'count'})
df_meta_summed = pd.merge(stack_guilds, df_meta, left_on = 'sample', right_on = 'SampleID', how='left')
df_meta_summed['long.name'] = df_meta_summed['sample'].apply(lambda x: x.replace('-','.'))
df_meta_summed.to_csv("guilds_02_read_greenhouse_meta_rhizo.csv", sep=",", header=True, index=False)

df_meta_summed.head()

# Roots

# Subsetting samples
df.columns.values
columns_sub = [ele for ele in df if 'Racine' in ele and 'Ctrl' not in ele]
df_abu_sub = df_abu[columns_sub]
print(df_abu_sub.shape)

# Summing guilds
df_sum = df_abu_sub.groupby(df_guilds['group']).sum()
df_sum.to_csv("guilds_02_read_greenhouse_roots.csv", sep=",", header=True, index=True)

# Merging with metadata
df_meta = pd.read_csv("/mnt/d/NRCan/2022_MicrobiomeAspenGreenhouse/DATA/PeuplierGH_jan2024/metadata_ITS.txt", sep="\t", header=0)
dt_sum = df_sum.T
stack_guilds = dt_sum.stack().reset_index().rename(columns={'level_0' : 'sample', 0 : 'count'})
df_meta_summed = pd.merge(stack_guilds, df_meta, left_on = 'sample', right_on = 'SampleID', how='left')
df_meta_summed['long.name'] = df_meta_summed['sample'].apply(lambda x: x.replace('-','.'))
df_meta_summed.to_csv("guilds_02_read_greenhouse_meta_roots.csv", sep=",", header=True, index=False)

df_meta_summed.head()
